[폴더 선택삭제] 버튼 및 [폴더 일괄삭제] 버튼 기능 구현 (ListView SelectedItem 바인딩)

+++ Bin_Obj_Delete_Project/ViewModels/MainVM.cs
@@ -17,7 +17,7 @@ namespace Bin_Obj_Delete_Project.ViewModels
         #region [프로퍼티]
 
         /// <summary>
-        /// [_buttonEnalbedOrNot]
+        /// [_delBtnEnabledOrNot]
         /// </summary>
         private bool _delBtnEnabledOrNot;
 
@@ -27,9 +27,24 @@ namespace Bin_Obj_Delete_Project.ViewModels
         private string _deleteFolderPath;
 
         /// <summary>
-        /// [_binObjToDelete]
+        /// [_deleteFolderInfo]
         /// </summary>
         private ObservableCollection<DeleteFolderInfo> _deleteFolderInfo;
+
+        /// <summary>
+        /// [_selectedCrFolder]
+        /// </summary>
+        private DeleteFolderInfo _selectedCrFolder;
+
+        /// <summary>
+        /// [_selectFolderInfo]
+        /// </summary>
+        private ObservableCollection<DeleteFolderInfo> _selectFolderInfo;
+
+        /// <summary>
+        /// [_activeFolderInfo]
+        /// </summary>
+        private ObservableCollection<DeleteFolderInfo> _activeFolderInfo;
 
         #endregion
 
@@ -38,7 +53,7 @@ namespace Bin_Obj_Delete_Project.ViewModels
         public event PropertyChangedEventHandler PropertyChanged;
 
         /// <summary>
-        /// [ButtonEnabledOrNot]
+        /// [IsDelBtnEnabledOrNot]
         /// </summary>
         public bool IsDelBtnEnabledOrNot
         {
@@ -61,7 +76,7 @@ namespace Bin_Obj_Delete_Project.ViewModels
         public string DeleteFolderPath
         {
             get => _deleteFolderPath;
-            private set
+            set
             {
                 if (_deleteFolderPath != value)
                 {
@@ -74,16 +89,72 @@ namespace Bin_Obj_Delete_Project.ViewModels
         }
 
         /// <summary>
-        /// [DelefeFolderInfo]
+        /// [DeleteFolderInfo]
         /// </summary>
-        public ObservableCollection<DeleteFolderInfo> DelefeFolderInfo
+        public ObservableCollection<DeleteFolderInfo> DeleteFolderInfo
         {
             get => _deleteFolderInfo;
-            private set
+            set
             {
                 if (_deleteFolderInfo != value)
                 {
                     _deleteFolderInfo = value;
+                    OnPropertyChanged();
+                }
+
+            }
+
+        }
+
+        /// <summary>
+        /// [SelectedCrFolder]
+        /// </summary>
+        public DeleteFolderInfo SelectedCrFolder
+        {
+            get => _selectedCrFolder;
+            set
+            {
+                if (_selectedCrFolder != value)
+                {
+                    _selectedCrFolder = value;
+                    OnPropertyChanged();
+                }
+
+            }
+
+        }
+
+        /// <summary>
+        /// [SelectFolderInfo]
+        /// </summary>
+        public ObservableCollection<DeleteFolderInfo> SelectFolderInfo
+        {
+            get => _selectFolderInfo;
+            set
+            {
+                if (_selectFolderInfo != value)
+                {
+                    _selectFolderInfo = value;
+                    OnPropertyChanged();
+                }
+
+            }
+
+        }
+
+        /// <summary>
+        /// [ActiveFolderInfo]
+        /// [폴더 선택삭제] 시 => [ActiveFolderInfo = SelectFolderInfo]
+        /// [폴더 일괄삭제] 시 => [ActiveFolderInfo = DeleteFolderInfo]
+        /// </summary>
+        public ObservableCollection<DeleteFolderInfo> ActiveFolderInfo
+        {
+            get => _activeFolderInfo;
+            set
+            {
+                if (_activeFolderInfo != value)
+                {
+                    _activeFolderInfo = value;
                     OnPropertyChanged();
                 }
 
@@ -101,29 +172,37 @@ namespace Bin_Obj_Delete_Project.ViewModels
         #region [ICommand]
 
         // 1. 폴더 불러오기
-        public ICommand LoadFolderCommand { get; set; }
+        public ICommand LoadingFolderCommand { get; set; }
 
-        // 2. bin, obj 폴더 삭제
-        public ICommand DelFoldersCommand { get; set; }
+        // 2. 폴더 선택삭제
+        public ICommand DelSelFoldersCommand { get; set; }
+
+        // 3. 폴더 일괄삭제
+        public ICommand DelAllFoldersCommand { get; set; }
 
         #endregion
-
 
         #region 생성자 (Initialize)
 
         public MainVM()
         {
-            LoadFolderCommand = new RelayCommand(LoadFolder);
+            LoadingFolderCommand = new RelayCommand(LoadingFolder);
+            _selectedCrFolder = new DeleteFolderInfo();
+            _selectFolderInfo = new ObservableCollection<DeleteFolderInfo>();
+            _activeFolderInfo = new ObservableCollection<DeleteFolderInfo>();
             _deleteFolderInfo = new ObservableCollection<DeleteFolderInfo>();
-            DelFoldersCommand = new RelayCommand(DelFolders);
+            DelSelFoldersCommand = new RelayCommand(DelSelFolders);
+            DelAllFoldersCommand = new RelayCommand(DelAllFolders);
         }
 
         #endregion
 
+        #region [버튼기능]
+
         /// <summary>
         /// 1. [폴더 불러오기] 기능
         /// </summary>
-        private void LoadFolder()
+        private void LoadingFolder()
         {
             CommonOpenFileDialog folderDialog = new CommonOpenFileDialog
             {
@@ -141,11 +220,12 @@ namespace Bin_Obj_Delete_Project.ViewModels
                 // [dirSubInfo] = 하위 폴더 리스트 정보
                 if (!string.IsNullOrEmpty(DeleteFolderPath))
                 {
+                    // [directoryInfo] 관련 참고해야 할 코드 (By. 박상훈 선임)
                     //var directoryInfo = new DirectoryInfo(DeleteFolderPath);
-                    //var directoryInfos = directoryInfo.GetFileSystemInfos("*", SearchOption.AllDirectories); // 내일 참고해서 코드 짜기
+                    //var directoryInfos = directoryInfo.GetFileSystemInfos("*", SearchOption.AllDirectories);
                     IEnumerable<string> directories = Directory.EnumerateDirectories(DeleteFolderPath, "*", SearchOption.AllDirectories);
-                    //.Where(dir => dir.EndsWith("bin") || dir.EndsWith("obj")); // 경로의 마지막 글자가 "bin"이거나 "obj"인 파일만 찾음.
-                    DelefeFolderInfo?.Clear(); // 컬렉션 초기화
+                    //.Where(dir => dir.EndsWith("bin") || dir.EndsWith("obj")); // 경로의 마지막 글자가 "bin"이거나 "obj"인 파일만 찾음!
+                    DeleteFolderInfo?.Clear(); // (전체) 컬렉션 초기화
                     // 파일 경로가 존재할 때,
                     if (directories != null)
                     {
@@ -155,12 +235,11 @@ namespace Bin_Obj_Delete_Project.ViewModels
                             // 해당 경로의 [폴더 이름]이 "bin" 또는 "obj"인 경우에만 포함
                             if (dirInfo.Name.Equals("bin") || dirInfo.Name.Equals("obj"))
                             {
-                                DelefeFolderInfo.Add(new DeleteFolderInfo
+                                DeleteFolderInfo.Add(new DeleteFolderInfo
                                 {
                                     DelFolderPath = dir,
                                     DelFolderName = dirInfo.Name,
                                     DelFolderCreationTime = dirInfo.CreationTime.ToString(),
-                                    DelFolderAccessedTime = dirInfo.LastAccessTime.ToString(),
                                     DelFolderModifiedTime = dirInfo.LastWriteTime.ToString(),
                                     DelFolderCategory = "파일 폴더",
                                     DelFolderSize = dir.Length.ToString() + " Byte"
@@ -177,16 +256,53 @@ namespace Bin_Obj_Delete_Project.ViewModels
                     }
 
                 }
-                // DeleteFolderInfo 정보 확인: 디버깅!
-                foreach (DeleteFolderInfo item in DelefeFolderInfo)
+                // DeleteFolderInfo 정보 확인: 디버깅으로 확인 가능
+                foreach (DeleteFolderInfo item in DeleteFolderInfo)
                 {
                     Console.WriteLine(item.DelFolderPath);
                     Console.WriteLine(item.DelFolderName);
                     Console.WriteLine(item.DelFolderCreationTime);
-                    Console.WriteLine(item.DelFolderAccessedTime);
                     Console.WriteLine(item.DelFolderModifiedTime);
                     Console.WriteLine(item.DelFolderCategory);
                     Console.WriteLine(item.DelFolderSize);
+                }
+                ActiveFolderInfo = DeleteFolderInfo; // [ActiveFolderInfo] 컬렉션에 [DeleteFolderInfo] 컬렉션을 할당
+            }
+
+        }
+
+        /// <summary>
+        /// 2. [폴더 선택삭제] 기능
+        /// </summary>
+        private void DelSelFolders()
+        {
+            IsDelBtnEnabledOrNot = false; // [폴더 선택삭제] 버튼 비활성화
+            if (!string.IsNullOrEmpty(DeleteFolderPath))
+            {
+                if (SelectFolderInfo?.Count > 0)
+                {
+                    // [SelectFolderInfo] 컬렉션 항목들 기반 새로운 List<DeleteFolderInfo> 객체를 생성함.
+                    List<DeleteFolderInfo> selectToDelete = new List<DeleteFolderInfo>(SelectFolderInfo);
+                    foreach (DeleteFolderInfo folder in selectToDelete)
+                    {
+                        string dir = folder.DelFolderPath;
+                        // [try ~ catch]문 활용 예외 처리!
+                        try
+                        {
+                            if (Directory.Exists(dir))
+                            {
+                                Directory.Delete(dir, true); // 지정된 디렉터리 및 해당 디렉터리의 하위 디렉터리 및 파일 삭제
+                                _ = ActiveFolderInfo.Remove(folder); // [ActiveFolderInfo] 컬렉션 => selectToDelete 항목 제외
+                            }
+                            IsDelBtnEnabledOrNot = true; // [폴더 선택삭제] 버튼 활성화
+                        }
+                        catch (Exception ex)
+                        {
+                            Console.WriteLine($"Error Deleting Folder... FolderPath: {dir} : {ex.Message}");
+                        }
+
+                    }
+
                 }
 
             }
@@ -194,36 +310,40 @@ namespace Bin_Obj_Delete_Project.ViewModels
         }
 
         /// <summary>
-        /// 2. [bin, obj 폴더 삭제] 기능
+        /// 3. [폴더 일괄삭제] 기능
         /// </summary>
-        private void DelFolders()
+        private void DelAllFolders()
         {
-            IsDelBtnEnabledOrNot = false; // [bin, obj 폴더 삭제] 버튼 비활성화
+            IsDelBtnEnabledOrNot = false; // [폴더 일괄삭제] 버튼 비활성화
             if (!string.IsNullOrEmpty(DeleteFolderPath))
             {
-                foreach (DeleteFolderInfo folder in DelefeFolderInfo)
+                if (DeleteFolderInfo?.Count > 0)
                 {
-                    string dir = folder.DelFolderPath;
-                    // [try ~ catch]문 활용 예외 처리!
-                    try
+                    foreach (DeleteFolderInfo folder in DeleteFolderInfo)
                     {
-                        if (Directory.Exists(dir))
+                        string dir = folder.DelFolderPath;
+                        // [try ~ catch]문 활용 예외 처리!
+                        try
                         {
-                            Directory.Delete(dir, true); // 지정된 디렉터리 및 해당 디렉터리의 하위 디렉터리 및 파일 삭제
+                            if (Directory.Exists(dir))
+                            {
+                                Directory.Delete(dir, true); // 지정된 디렉터리 및 해당 디렉터리의 하위 디렉터리 및 파일 삭제
+                            }
+                            IsDelBtnEnabledOrNot = true; // [폴더 일괄삭제] 버튼 활성화
+                        }
+                        catch (Exception ex)
+                        {
+                            Console.WriteLine($"Error Deleting Folder... FolderPath: {dir} : {ex.Message}");
                         }
 
                     }
-                    catch (Exception ex)
-                    {
-                        Console.WriteLine($"Error Deleting Folder... FolderPath: {dir} : {ex.Message}");
-                    }
-
+                    DeleteFolderInfo?.Clear(); // (전체) 컬렉션 초기화
                 }
 
             }
-            DelefeFolderInfo?.Clear(); // 컬렉션 초기화
+
         }
 
     }
-
+    #endregion
 }
